backend/requirements.txt
Django~=3.1.4
psycopg2~=2.8.6
channels~=3.0.3
channels_redis~=3.2.0
djangorestframework~=3.12.2
djangorestframework-jwt~=1.11.0
django-cors-headers~=3.7.0
pillow~=8.1.2
uuid~=1.30